// src/shared/api/generated/model/recentPlace.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 아이랑 - 홈 화면 API
 * OpenAPI spec version: 0.0.1
 */
import type { Region } from './region';

export interface RecentPlace {
  name: string;
  id: string;
  region: Region;
  /** @nullable */
  imgUrl: string | null;
  isStrollerRental: boolean;
  isPark: boolean;
  isToilet: boolean;
  /** 생태관광지면 true, 축제면 false */
  isEcoTourism: boolean;
  /** @nullable */
  operatingHours: string | null;
  /** @nullable */
  price: string | null;
  latitude: number;
  longitude: number;
  /**
     * 축제일 때만 (YYYYMMDD)
     * @nullable
     */
  startAt?: string | null;
  /**
     * 축제일 때만 (YYYYMMDD)
     * @nullable
     */
  endAt?: string | null;
}